Move queries and filters to utils/queries
