+++ admin-src/index.js
@@ -7,50 +7,7 @@ import { LoginScreen } from './LoginScreen'
 import { Button, Row } from 'react-bootstrap'
 import { Tab, Tabs, TabList, TabPanel } from 'react-tabs'
 import 'react-tabs/style/react-tabs.css'
-
-const STALE_STATUS_AFTER_DAYS = 10
-
-const unansweredQuery = (onSnapshot) => {
-  return firebase
-    .firestore()
-    .collectionGroup('requests')
-    .where('amountOfAnswers', '==', 0)
-    .onSnapshot(onSnapshot)
-}
-const dateDaysAgo = (daysAgo) => {
-  const date = new Date(Date.now())
-  date.setDate(date.getDate() - daysAgo)
-  return date
-}
-const filterBySolved = (diagnose) => diagnose.solved
-const filterByNotSolved = (diagnose) => !diagnose.solved
-const filterByLastActivity = (diagnose) => (diagnose.updatedAt !== undefined) && diagnose.updatedAt.toDate() >= dateDaysAgo(STALE_STATUS_AFTER_DAYS)
-const filterByAmountOfAnswers = (diagnose) => diagnose.amountOfAnswers > 0
-const filterStale = (diagnose) => !filterByLastActivity(diagnose) && filterByNotSolved(diagnose) && filterByAmountOfAnswers(diagnose)
-const filterInDiscussion = (diagnose) => filterByLastActivity(diagnose) && filterByNotSolved(diagnose) && filterByAmountOfAnswers(diagnose)
-
-const inDiscussionQuery = (onSnapshot) => {
-  return firebase
-    .firestore()
-    .collectionGroup('requests')
-    .where('amountOfAnswers', '>', 0)
-    .onSnapshot(async (snapshot) => onSnapshot(snapshot, filterInDiscussion))
-}
-
-const staleQuery = (onSnapshot) => {
-  return firebase
-    .firestore()
-    .collectionGroup('requests')
-    .where('amountOfAnswers', '>', 0)
-    .onSnapshot(async (snapshot) => onSnapshot(snapshot, filterStale))
-}
-
-const solvedQuery = (onSnapshot) => {
-  return firebase
-    .firestore()
-    .collectionGroup('requests')
-    .onSnapshot(async (snapshot) => onSnapshot(snapshot, filterBySolved))
-}
+import { unansweredQuery, inDiscussionQuery, staleQuery, solvedQuery } from './utils/queries'
 
 const App = () => {
   const [userLoggedIn, setUserLoggedIn] = useState(false)
